{LYN5384} splitting Blast builder and processor Python scripts
splitting the asset builder and asset processor Python scripts for the Blast processor
This fixes a mulitiple build issue found while developing other scripts
